refactor: restructure layout components and base layout
