+++ layout/components/Footer/IntroList.jsx
